==> src/2_cat_catEncoder.py <==
# -*- coding: utf-8 -*-
"""
@author: Alex Tselikov
"""

from mers_utils import *
from sklearn.preprocessing import LabelEncoder
import warnings
warnings.filterwarnings("ignore", category=DeprecationWarning) 
import argparse

def catEncoder(series, nCats=10):
            return dict([(x, i) for i, x in enumerate(series.value_counts().index[:nCats])])

if __name__ == "__main__":
    t0 = time()
    parser = argparse.ArgumentParser(description='create pca features',
                                     formatter_class=argparse.ArgumentDefaultsHelpFormatter)
    parser.add_argument('ncats', type=int, help='training data')
    parser.add_argument('input', type=str, help='training data')
    args = parser.parse_args()
    
    iinput=args.input
    ncats=args.ncats

    ## READ
    train = pd.read_csv(FFOLD+iinput+'.train.csv')
    test = pd.read_csv(FFOLD+iinput+'.test.csv')
    df = pd.concat([train, test])

    ## PROCESS
    cat_cols=train.select_dtypes(include=['object']).columns
    df = df[cat_cols]
    for col in cat_cols:
        enc = catEncoder(df[col], ncats)
        df[col] = df[col].map(enc)
    df = df.fillna(-1)
    ## SAVE
    np.savetxt(FFOLD+'2cat_catEncoded'+str(ncats)+'_'+iinput+'.train.csv',df[:len(train)].values, fmt='%10.0f')
    np.savetxt(FFOLD+'2cat_catEncoded'+str(ncats)+'_'+iinput+'.test.csv',df[len(train):].values, fmt='%10.0f')
    print ('DONE in %s min' % np.round((time() - t0) / 60. ,3))
